[WiFi] Fix bogus RTC settings loaded to connect to WiFi

+++ src/src/Helpers/Networking.cpp
@@ -211,7 +211,7 @@ void checkUDP()
       std::vector<char> packetBuffer;
       packetBuffer.resize(packetSize + 1);
 
-      if (packetBuffer.size() >= packetSize) {
+      if (packetBuffer.size() >= static_cast<size_t>(packetSize)) {
         memset(&packetBuffer[0], 0, packetSize + 1);
         int len = portUDP.read(&packetBuffer[0], packetSize);
